work/requirements.txt
fastapi[all]==0.74.1

databases[postgresql]==0.5.5
aerich==0.6.3
tortoise-orm[asyncpg]==0.19.0